requirements.txt
amqp==1.4.9
anyjson==0.3.3
astroid==2.3.2
autopep8==1.4.4
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.9.11
defusedxml==0.6.0
diff-match-patch==20181111
Django==1.11.18
django-bootstrap-ui==0.5.1
django-celery==3.3.1
django-cors-headers==3.1.1
django-debug-toolbar==2.0
django-import-export==1.2.0
django-mdeditor==0.1.16
django-redis==4.10.0
django-rest-framework==0.1.0
django-stdimage==5.0.1
django-tag-parser==2.1
django-treebeard==4.3
djangorestframework==3.10.3
docopt==0.6.2
dominate==2.1.17
et-xmlfile==1.0.1
idna==2.8
isort==4.3.21
jdcal==1.4.1
kombu==3.0.37
lazy-object-proxy==1.4.2
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.0
openpyxl==3.0.0
Pillow==6.2.1
progressbar2==3.47.0
pycodestyle==2.5.0
pylint==2.4.3
PyMySQL==0.9.3
python-utils==2.3.0
pytz==2019.3
PyYAML==5.1.2
redis==3.3.11
simpleui==3.2
six==1.12.0
sqlparse==0.3.0
tablib==0.14.0
typed-ast==1.4.0
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
